update to 17.04

+++ src/app/admin/timetable/timetable.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Inject } from "@angular/core";
+import { ActivatedRoute } from "@angular/router";
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from "@angular/material";
-import TableService, { Subject, Group, TimeEntity } from "./timetable.service";
+import TableService, { Subject, Group, TableItem } from "./timetable.service";
 import {
   FormControl,
   FormGroup,
@@ -8,8 +9,10 @@ import {
   ReactiveFormsModule
 } from "@angular/forms";
 import { TimeTableModal } from "./timetable-modal/timetable-modal.component";
+import { TimetableDeleteConfirmComponent } from "./timetable-delete-confirm/timetable-delete-confirm.component";
+import { ResponseMessageComponent } from "../../shared/response-message/response-message.component";
 
-interface TimeEntityTable extends TimeEntity {
+interface TableItemModified extends TableItem {
   subject_name: string;
   group_name: string;
 }
@@ -21,56 +24,75 @@ interface TimeEntityTable extends TimeEntity {
   providers: [TableService]
 })
 export class TimetableComponent implements OnInit {
-  table: TimeEntityTable[] = [];
+  table: TableItemModified[] = [];
   subjects: Subject[] = [];
   groups: Group[] = [];
 
   groupsMap: Map<string, string> = new Map();
   subjectsMap: Map<string, string> = new Map();
 
-  constructor(public tableService: TableService, public dialog: MatDialog) {
-  /** 
-   *  login is FOR TEST ONLY
-   * TODO: REMOVE before prod 
-   */
-    tableService.login().subscribe(() => {
-      tableService.getTable().subscribe(
-        async data => {
-          await tableService
-            .getGroups()
-            .toPromise()
-            .then(groups => {
-              this.groups = groups;
+  constructor(
+    public tableService: TableService,
+    public dialog: MatDialog,
+    private route: ActivatedRoute
+  ) {
+    const onSuccess = async data => {
+      await tableService
+        .getGroups()
+        .toPromise()
+        .then((groups: Group[]) => {
+          this.groups = groups;
 
-              groups.forEach(group =>
-                this.groupsMap.set(group.group_id, group.group_name)
-              );
-            });
+          groups.forEach(group =>
+            this.groupsMap.set(group.group_id, group.group_name)
+          );
+        });
 
-          await tableService
-            .getSubjects()
-            .toPromise()
-            .then(subjects => {
-              this.subjects = subjects;
+      await tableService
+        .getSubjects()
+        .toPromise()
+        .then((subjects: Subject[]) => {
+          this.subjects = subjects;
 
-              subjects.forEach(subject =>
-                this.subjectsMap.set(subject.subject_id, subject.subject_name)
-              );
-            });
+          subjects.forEach(subject =>
+            this.subjectsMap.set(subject.subject_id, subject.subject_name)
+          );
+        });
 
-          this.table = data.map((timeEntity: TimeEntity): TimeEntityTable => {
-            return Object.assign({}, timeEntity, {
-              group_name: this.groupsMap.get(timeEntity.group_id),
-              subject_name: this.subjectsMap.get(timeEntity.subject_id)
-            });
-          });
+      // If no records found
+      if (!Array.isArray(data)) {
+        this.table = [];
+        this.dialog.open(ResponseMessageComponent, {
+          width: "400px",
+          data: {
+            message: "За даним запитом розкладу не знайдено"
+          }
+        });
+        return;
+      }
 
-          console.log(this.table);
-        },
-        err => {
-          this.table = [];
-        }
-      );
+      this.table = data.map((timeEntity: TableItem): TableItemModified => {
+        return Object.assign({}, timeEntity, {
+          group_name: this.groupsMap.get(timeEntity.group_id),
+          subject_name: this.subjectsMap.get(timeEntity.subject_id)
+        });
+      });
+    };
+
+    const onError = () => (this.table = []);
+
+    this.route.queryParams.subscribe(params => {
+      if (params.subjectId) {
+        tableService
+          .getTableBySubjectId(params.subjectId)
+          .subscribe(onSuccess, onError);
+      } else if (params.groupId) {
+        tableService
+          .getTableByGroupId(params.groupId)
+          .subscribe(onSuccess, onError);
+      } else {
+        tableService.getTable().subscribe(onSuccess, onError);
+      }
     });
   }
 
@@ -78,24 +100,44 @@ export class TimetableComponent implements OnInit {
    * Delete table item
    * id of table item to delete
    */
-  onDelete(timeEntity: TimeEntityTable) {
-    this.tableService.deleteTableItem(timeEntity.timetable_id).subscribe(
-      response => {
-        return this.table.splice(this.table.indexOf(timeEntity), 1);
-      },
-      err => {
-        console.error("err:", err);
-        alert("Сервер вернув помилку. Спробуйте пізніше...");
-      }
-    );
-  }
+  onDelete(timeEntity: TableItemModified) {
+    let dialogRef = this.dialog.open(TimetableDeleteConfirmComponent, {
+      width: "400px"
+    });
 
+    dialogRef.afterClosed().subscribe((response: string) => {
+      if (response) {
+        if (response === "ok") {
+          this.tableService.deleteTableItem(timeEntity.timetable_id).subscribe(
+            response => {
+              this.table.splice(this.table.indexOf(timeEntity), 1);
+              this.dialog.open(ResponseMessageComponent, {
+                width: "400px",
+                data: {
+                  message: "Розклад успішно видалено!"
+                }
+              });
+            },
+            err => {
+              console.error("err:", err);
+              this.dialog.open(ResponseMessageComponent, {
+                width: "400px",
+                data: {
+                  message: "Виникла помилка при видаленні розкладу!"
+                }
+              });
+            }
+          );
+        }
+      }
+    });
+  }
   /**
    * Opens dialog to create table entity or edit it
    * table item we want to edit
    * if not presented we open modal to add new entity
    */
-  openDialog(editData: TimeEntityTable): void {
+  openDialog(tableItem: TableItemModified): void {
     let dialogRef = this.dialog.open(TimeTableModal, {
       width: "800px",
       data: {
@@ -104,14 +146,21 @@ export class TimetableComponent implements OnInit {
         groups: this.groups,
         groupsMap: this.groupsMap,
         subjectsMap: this.subjectsMap,
-        editData
+        tableItem
       }
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      console.log("The dialog was closed, result:" + result);
+      if(result){
+         this.dialog.open(ResponseMessageComponent, {
+        width: "400px",
+        data: {
+          message: result
+        }
+      });
+      }
     });
   }
-
+  
   ngOnInit() {}
 }
